fix: refactor package tattooclientprofile
