+++ app/src/main/res/layout/fragment_tattoo_client_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.tattooclientprofile.TattooClientProfileFragment">
+    tools:context=".ui.profile.tattooclient.TattooClientProfileFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/linearLayout2"
