+++ app/src/main/res/navigation/user_page_navigation_graph.xml
@@ -12,7 +12,7 @@
         tools:layout="@layout/fragment_home_user" />
     <fragment
         android:id="@+id/clientUserProfileFragment"
-        android:name="br.com.connectattoo.ui.tattooclientprofile.TattooClientProfileFragment"
+        android:name="br.com.connectattoo.ui.profile.tattooclient.TattooClientProfileFragment"
         android:label="fragment_client_user_profile"
         tools:layout="@layout/fragment_tattoo_client_profile" />
 </navigation>
